Use macro instead of quotes for Quarkus output
This commit updates the Quarkus command output block to use the macros
attribute and a macro for quotes substitution for the two indexed
numbers in the output.

The motivation for this change is that currently the ascii-art
spelling out Quarkus is not displayed properly. With this change it will
be diplayed properly and the quotes in the output will be quoted as the
currently are.